Update firewalld rules as expected
Currently we were exposing '192.168.126.0/24' over trusted zone
which means it's publicly exposed but we just need it for the libvirt
communication inside the host. zone `dmz` is the internal one and can
be used for this purpose. Also to forward the request from `192.168.126.0/24`
to `192.168.122.1` it is easier to add using firewalld cmd with specific
zone instead of using `iptables` one.